+++ src/topology.c
@@ -948,7 +948,7 @@ int igraph_isoclass_create(igraph_t *graph, igraph_integer_t size,
       classedges=igraph_i_classedges_3;
 
       if (number < 0 || 
-	  number >= sizeof(igraph_i_isographs_3)/sizeof(unsigned int)) {
+	  number >= (int)(sizeof(igraph_i_isographs_3)/sizeof(unsigned int))){
 	IGRAPH_ERROR("`number' invalid, cannot create graph", IGRAPH_EINVAL);
       }
 
@@ -958,7 +958,7 @@ int igraph_isoclass_create(igraph_t *graph, igraph_integer_t size,
       classedges=igraph_i_classedges_4;
 
       if (number < 0 ||
-	  number >= sizeof(igraph_i_isographs_4)/sizeof(unsigned int)) {
+	  number >= (int)(sizeof(igraph_i_isographs_4)/sizeof(unsigned int))){
 	IGRAPH_ERROR("`number' invalid, cannot create graph", IGRAPH_EINVAL);
       }
 
@@ -970,7 +970,8 @@ int igraph_isoclass_create(igraph_t *graph, igraph_integer_t size,
       classedges=igraph_i_classedges_3u;
 
       if (number < 0 ||
-	  number >= sizeof(igraph_i_isographs_3u)/sizeof(unsigned int)) {
+	  number >= (int)(sizeof(igraph_i_isographs_3u)/
+			  sizeof(unsigned int))){
 	IGRAPH_ERROR("`number' invalid, cannot create graph", IGRAPH_EINVAL);
       }
 
@@ -980,7 +981,8 @@ int igraph_isoclass_create(igraph_t *graph, igraph_integer_t size,
       classedges=igraph_i_classedges_4u;
 
       if (number < 0 ||
-	  number >= sizeof(igraph_i_isographs_4u)/sizeof(unsigned int)) {
+	  number >= (int)(sizeof(igraph_i_isographs_4u)/
+			  sizeof(unsigned int))) {
 	IGRAPH_ERROR("`number' invalid, cannot create graph", IGRAPH_EINVAL);
       }
